tracking relaciones orden-shopping_cart-usuarios-direccion

+++ database/migrations/2017_07_21_051518_create_orders_table.php
@@ -56,7 +56,7 @@ class CreateOrdersTable extends Migration
 
             $table->string('guide_numer')->nullable(); 
 
-            $table->integer("total"); 
+            $table->decimal("total",8,2); 
 
             $table->timestamps();
         });
